overload function rm and fix some small bugs

+++ src/feature/gen_path_str.h
@@ -9,7 +9,7 @@ string gen_path_str(){
     string s = "~";
     string n = User_list.at(current_user).home_dir.u8string();
     string m = fs_current_path.u8string();
-    for(int i=0;i<n.length();i++){
+    for(size_t i=0;i<n.length();i++){
         if(n[i] != m[i]){
             s = m;
             return s;
